feat(client): add support for sacaria transactions
Add support for sacaria transactions by creating a new type of transaction
called SacariaTransactionModel. Also, create a new form component to
credit sacaria to a client's account. The form is displayed in a modal
that can be opened from the sacaria account table. The form data is sent
to the server using the createSacariaService function.