feat: add required prop to input fields that are required
This commit adds the required prop to input fields that are required in the following files:
- src/domain/client/view/Client/Accounts/Checking/components/Form/index.tsx
- src/domain/client/view/Client/Accounts/Sacaria/components/Form/index.tsx
- src/shared/components/inputs/ControllerField.tsx
- src/shared/components/inputs/RHFCurrencyInput/index.tsx

feat: add CustomInput prop to ControllerField

This commit adds the CustomInput prop to the ControllerField component in the shared/components/inputs/ControllerField.tsx file. This prop allows the user to pass a custom input component to the ControllerField component.

fix: change label text in SacariaFormView

This commit changes the label text in the SacariaFormView component in the src/domain/client/view/Client/Accounts/Sacaria/components/Form/index.tsx file to include an asterisk for required fields.

fix