feat(ClientBalances): add accordion to show/hide details of coffee balances
This commit adds an accordion component to the ClientBalances view to
show/hide details of coffee balances. The accordion has a single item
that displays the balance of coffee bags and, when expanded, shows the
balance of the "Duro" coffee type. This change improves the user
experience by allowing the user to see more information only when
needed.