feat(coffee): add functionality to buy coffee
Add functionality to buy coffee by creating a new service function to handle
the post request to the server. Also, add a new view to display a form to buy
coffee and handle the form submission by calling the new service function.
The form calculates the total value of the coffee based on the weight and the
number of bags. The weight is calculated based on the weight per bag constant
defined in the GlobalConfig file. The service function receives the clientId
to identify the client who is buying the coffee.