feat(Checking/TransactionsTable/Row.tsx): add expandable description column to show full transaction description on click
The description column in the TransactionsTable component was modified to be expandable. Now, when the user clicks on the description cell, the full transaction description is shown. This was done to improve the user experience and allow the user to see the full description of the transaction without having to navigate to another page. The Collapse component from Chakra UI was used to achieve this functionality.